✨ Apply react lazy on preview table to reduce bundle size
